requirements.txt
asgiref==3.3.1
Django==3.1.4
django-uuslug==1.2.0
djangorestframework==3.12.2
Pillow==8.1.0
python-slugify==4.0.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
